requirements.txt
channels==2.1.4
Django==2.1.3
djangorestframework==3.9.0
gunicorn==19.9.0
PyMySQL==0.9.2
python-jenkins==1.4.0
Twisted==18.9.0
uWSGI==2.0.17.1
